# docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "EventAckIn": {
        "properties": {
          "cursor": {
            "minimum": 0.0,
            "title": "Cursor",
            "type": "integer"
          },
          "stream": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Stream",
            "type": "string"
          },
          "subscriber_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Subscriber Id",
            "type": "string"
          }
        },
        "required": [
          "stream",
          "subscriber_id",
          "cursor"
        ],
        "title": "EventAckIn",
        "type": "object"
      },
      "EventAckOut": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "stored_cursor": {
            "title": "Stored Cursor",
            "type": "integer"
          }
        },
        "required": [
          "ok",
          "stored_cursor",
          "status"
        ],
        "title": "EventAckOut",
        "type": "object"
      },
      "EventEmitIn": {
        "properties": {
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "trace_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "type": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "payload"
        ],
        "title": "EventEmitIn",
        "type": "object"
      },
      "EventEmitOut": {
        "properties": {
          "event_id": {
            "title": "Event Id",
            "type": "integer"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "event_id"
        ],
        "title": "EventEmitOut",
        "type": "object"
      },
      "EventItemOut": {
        "properties": {
          "cursor": {
            "title": "Cursor",
            "type": "integer"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "cursor",
          "ts",
          "type",
          "payload"
        ],
        "title": "EventItemOut",
        "type": "object"
      },
      "EventReadOut": {
        "properties": {
          "cursor": {
            "title": "Cursor",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "integer"
          },
          "stream": {
            "title": "Stream",
            "type": "string"
          }
        },
        "required": [
          "stream",
          "cursor",
          "next_cursor",
          "items"
        ],
        "title": "EventReadOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Device-Token",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "operationId": "read_events_api_v1_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "stream",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Stream",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Read Events",
        "tags": [
          "events",
          "events"
        ]
      }
    },
    "/api/v1/events/ack": {
      "post": {
        "operationId": "ack_events_api_v1_events_ack_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventAckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventAckOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Ack Events",
        "tags": [
          "events",
          "events"
        ]
      }
    },
    "/api/v1/events/emit": {
      "post": {
        "operationId": "emit_event_api_v1_events_emit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventEmitIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventEmitOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Emit Event",
        "tags": [
          "events",
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "operationId": "get_event_api_v1_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Event",
        "tags": [
          "events",
          "events"
        ]
      }
    },
...
  }
}